requirements.txt
Flask
python-dotenv
supabase
pypdf
google-genai
